Removed ESP8266 manual
Decided to stick to Bluetooth. It's a serial interface already, there's no need to create any interface like with ESP8266. Also, we don't need to trust external (broken) libraries.

The connection problem I was facing with Bluetooth wasn't actually a connection problem. Even thought macOS says the device is "disconnected" after about 10 seconds you pair it with the laptop, it's connected. It will only show as "connected" when a serial communication with the device has started. I tested it initially with Arduino IDE's serial monitor. Opening it "connects" the HC05. Closing it "disconnects it". If I'd known this before, I could have saved about 4 days of wasted work with the ESP8266 WiFi module.